Assignment_3/django-chatter-terminal/requirements.txt
Django==4.2.14
chatterbot2==1.1.0a7
chatterbot_corpus==1.2.0
SQLAlchemy
python-dateutil==2.8.2
